deps: update to ES Java client 8.15.5 in Optimize

+++ optimize/util/optimize-commons/src/main/java/io/camunda/optimize/service/util/configuration/db/DatabaseIndex.java
@@ -23,7 +23,7 @@ public class DatabaseIndex {
   private String refreshInterval;
 
   @JsonProperty("nested_documents_limit")
-  private Integer nestedDocumentsLimit;
+  private Long nestedDocumentsLimit;
 
   private String prefix;
 }
